Add time stamp to the log
The log file only contains the result of the call to the executable. Without
a time stamp it is extremely difficult to correlate the infromation from
instance-billing-flavor check with other events on the system or in the CSP
framework. Configuring logging to write the time stamp to the file addresses
this issue.