Improve flavor determination stability
In order to reliably determine the flavor of an instance, PAYG or BYOS, we
need to contact the update infrastructure. However the server may be very busy
and as such we might not get a response in the expected time. We will first
retry after a short wait period of 2 seconds and if we do still not get
and answer we report the status from the last successful determination of
the flavor. This opens a small window that an instance that was switched
between flavors will be misreported but overall the reliability of the check
is improved.