Make the function that checks the flavor reusable
Previously the function check_payg_byos which is the entry point to determine
the instance flavor exited using the sys.exit() Python function. This made
the function not reusable in other code bases that may need to determine the
flavor. Change the behavior of the function to return the flavor indicator
as a string and return a code indicating the flavor.