Avoid reaching out to the same server more than necessary
When we collect the IP address(es) from the hosts file for the
registration server the hosts file may have entries for the registry
and the repository service. 2 DNS entries pointing to the same IP address.
At present we will collect the IP address for both DNS names and then try
to verify the flavor of the instance against that same IP in 2 different
retry loops. Do not try to access a server more than once for a retry loop.